1-linkedin-post-generator/requirements.txt
crewai>=0.140.0
groq>=0.29.0
python-dotenv>=1.0.0